Fixed SwiftLint error and reorganized Task folder structure
